adjustments in mobile lines Banner
