+++ src/components/Home/Banner/LineTwo.tsx
@@ -19,12 +19,18 @@ const LineTwo: FC = () => {
           attributeName="d"
           values=" 
           M-65.9241 0.292397C-49.1524 32.6359 382.436 608.119 911.525 345.804C1440.61 83.4881 1779.84 97.1513 2026.63 255.206; 
-          M-65.9241 80.992397C-49.1524 100.6359 282.436 580.119 911.525 300.804C1440.61 100.4881 1779.84 400.1513 2026.63 455.206;
+          M-500.9241 500.292397C-49.1524 -200.6359 382.436 500.119 911.525 370.804C1440.61 200.4881 1779.84 -501.1513 2026.63 255.206;
           M-65.9241 0.292397C-49.1524 32.6359 382.436 608.119 911.525 345.804C1440.61 83.4881 1779.84 97.1513 2026.63 255.206"
           dur="6s"
           repeatCount="indefinite"
         />
       </path>
+      {/* <path
+        d="M-500.9241 600.292397C-49.1524 -200.6359 382.436 500.119 911.525 370.804C1440.61 200.4881 1779.84 -501.1513 2026.63 255.206"
+        stroke="#068A8F"
+        strokeWidth="5.11256"
+        strokeMiterlimit="10"
+      ></path> */}
     </svg>
   )
 }
